Rename DoubleClickHandler to 'TreDoubleClickListener' and move it to 'listener' package
