+++ src/main/java/com/redhat/devtools/intellij/tektoncd/WindowToolFactory.java
@@ -24,9 +24,8 @@ import com.intellij.ui.tree.AsyncTreeModel;
 import com.intellij.ui.tree.StructureTreeModel;
 import com.intellij.ui.treeStructure.Tree;
 import com.redhat.devtools.intellij.common.tree.MutableModelSynchronizer;
-import com.redhat.devtools.intellij.tektoncd.tree.DoubleClickHandler;
+import com.redhat.devtools.intellij.tektoncd.listener.TreeDoubleClickListener;
 import com.redhat.devtools.intellij.tektoncd.tree.TektonTreeStructure;
-import com.redhat.devtools.intellij.tektoncd.utils.TreeHelper;
 import org.jetbrains.annotations.NotNull;
 
 import java.lang.reflect.Constructor;
@@ -46,7 +45,7 @@ public class WindowToolFactory implements ToolWindowFactory {
             tree.setCellRenderer(new NodeRenderer());
             PopupHandler.installPopupHandler(tree, "com.redhat.devtools.intellij.tektoncd.tree", ActionPlaces.UNKNOWN);
             toolWindow.getContentManager().addContent(contentFactory.createContent(new JBScrollPane(tree), "", false));
-            DoubleClickHandler.install(tree, TreeHelper::openTektonResourceInEditor);
+           new TreeDoubleClickListener(tree);
         } catch (IllegalAccessException | InvocationTargetException | InstantiationException | NoSuchMethodException e) {
             throw new RuntimeException((e));
         }
